Server clean up and removal of temporary frontend boilerplate
