+++ .gitignore
@@ -1,2 +1,2 @@
 .env
-/Users/noorabid/ChapterRewind/(historical) old-flask-version
+node_modules
