--- views/home.ejs
@@ -1,9 +1,0 @@
-<%- layout('layouts/boilerplate') %>
-
-<div class="container">
-  <div class="inner-content">
-    <h1>ChapterRewind</h1>
-    <br />
-    <button><a href="/new">Start Search</a></button>
-  </div>
-</div>
