--- views/layouts/boilerplate.ejs
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Chapter Rewind</title>
-    <link rel="stylesheet" href="/styles.css" />
-  </head>
-  <body>
-    <%- body %>
-  </body>
-</html>
