--- views/new.ejs
@@ -1,24 +1,0 @@
-<%- layout('layouts/boilerplate') %>
-
-<div class="container">
-  <div class="inner-content">
-    <form id="book-search" action="/videos" method="POST">
-      <label for="title">Search By Title</label>
-      <input
-        type="text"
-        name="title"
-        id="title"
-        placeholder="Enter book title"
-      />
-      <label for="author">Search By Author</label>
-      <input
-        type="text"
-        name="author"
-        id="author"
-        placeholder="Enter author name"
-      />
-      <button>Search</button>
-    </form>
-    <Button><a href="/">Back to home</a></Button>
-  </div>
-</div>
