--- views/videos.ejs
@@ -1,25 +1,0 @@
-<%- layout("layouts/boilerplate")%> <% if (videos.length >0) { %>
-
-<div class="container video">
-  <div class="inner-content">
-    <ul>
-      <button><a href="/">Back to home</a></button>
-      <% videos.forEach(video =>{ %>
-      <li style="margin-bottom: 20px">
-        <a
-          href="https://youtube.com/watch?v=<%= video.id.videoId %>"
-          target="_blank"
-        >
-          <img
-            src="<%= video.snippet.thumbnails.medium.url %>"
-            alt="<%= video.snippet.title %>"
-          />
-          <p><%= video.snippet.title %></p>
-        </a>
-      </li>
-      <% })} else { %>
-      <p>No video found</p>
-      <% } %>
-    </ul>
-  </div>
-</div>
